requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
requests
Flask-Login
python-dotenv
